refactor: App에서 관리하는 카드 정보 훅으로 분리
